Remaking Stats Pages With Charts
This revamps the stat pages pretty heavily. Instead of listing out comic data, it's now rendering it via chart.

The client is making one call to the server at the moment to get the data and the majority of work is being done clientside to bucket and format the data. So far this seems pretty fast, but if move work needed to be moved over to the server it could be.

So far there's only 2 charts comparing gender between Marvel & DC, but it could easily (and probably will be) expanded on.