Re-doing profile pages
The profile pages were pretty hideous. This commit redoes a lot of that to make it prettier. It also adds in a very basic chart to make the profile page a bit more interesting.

Overall this makes the favorites much more prominent. Makes the marvel/dc gif a lot smaller and less important. Now much more responsive and fluid as well. Just overall much better