Adding saved messages
This makes the messages be saved when they are sent.

Additionally on page load, the chatting app with now display all the messages sent from the last 30 mins.

This also abstracts out a bit of code from the server file relating to the socket and moves it into it's own controller