Making chat and autocomplete play nicer
The chat and autocomplete functionality were playing extremely un-nicely. A lot of styling and hoisting issues causing the two to play really really unfriendly.

I had to change up some of the styling to get the chat where I wanted so the autocomplete results weren't covered. Additionally consolidated the chat.js and autocomplete.js files and make a comment relating to the important of the ordering in that single file.